version 1.3.6 DRY clean up
