+++ lib/glad.js
@@ -4,10 +4,10 @@ import chalk from 'chalk'
 
 import { GLADContext } from './gladContext.js'
 import { Graph } from './models/graph.js'
-import { DotParser } from './parsers/parseDot.js'
-import { FlutterDartParser } from './parsers/parseFlutterDart.js'
-import { JsTsParser } from './parsers/parseJsTs.js'
-import { SwiftParser } from './parsers/parseSwift.js'
+import { DotParser } from './parsers/parserDot.js'
+import { FlutterDartParser } from './parsers/parserFlutterDart.js'
+import { JsTsParser } from './parsers/parserJsTs.js'
+import { SwiftParser } from './parsers/parserSwift.js'
 
 /**
  * @typedef Options
