+++ lib/parsers/parserDot.js
@@ -2,6 +2,7 @@ import fs from 'fs'
 
 import { GLADContext } from '../gladContext.js'
 import { Container } from '../models/container.js'
+import { matchesExcludePattern } from '../util/strings.js'
 
 /**
  * DOT file parser for GLAD
@@ -52,7 +53,7 @@ export class DotParser {
     let filteredNodes = parsed.nodes
     if (this.context.options.exclude) {
       filteredNodes = parsed.nodes.filter(nodeName => {
-        const shouldExclude = this.matchesExcludePattern(nodeName, this.context.options.exclude)
+        const shouldExclude = matchesExcludePattern(nodeName, this.context.options.exclude)
         if (!this.context.options.silent && shouldExclude) {
           console.info(`Excluding node: ${nodeName}`)
         }
@@ -93,34 +94,6 @@ export class DotParser {
       if (leaf) {
         leaf.name = nodeName.split('/').pop()
       }
-    })
-  }
-
-  /**
-   * Check if a file path matches any exclude pattern
-   * @param {string} filePath - Relative file path to check
-   * @param {string|Array<string>} excludePatterns - Glob pattern(s) to match against
-   * @returns {boolean} True if the file should be excluded
-   */
-  matchesExcludePattern (filePath, excludePatterns) {
-    if (!Array.isArray(excludePatterns)) {
-      excludePatterns = excludePatterns ? [excludePatterns] : []
-    }
-
-    return excludePatterns.some(pattern => {
-      // Simple glob matching implementation
-      // First escape special regex characters including *
-      let regexPattern = pattern.replace(/[.+^${}()|[\]\\*?/]/g, '\\$&')
-
-      // Then convert escaped glob patterns to regex
-      regexPattern = regexPattern
-        .replace(/\\\*\\\*\\\//g, '(.*/)?')  // **/ matches optional path (escaped **/)
-        .replace(/\\\*\\\*/g, '.*')         // ** matches any path (escaped **)
-        .replace(/\\\*/g, '.*')             // * matches any characters including / (escaped *)
-        .replace(/\\\?/g, '[^/]')           // ? matches any single character except / (escaped ?)
-
-      const regex = new RegExp('^' + regexPattern + '$')
-      return regex.test(filePath)
     })
   }
