+++ lib/parsers/parserJsTs.js
@@ -1,20 +1,13 @@
 import * as TSMorph from 'ts-morph'
 
-import { GLADContext } from '../gladContext.js'
 import { Container } from '../models/container.js'
-import { stringIsEmpty } from '../util/strings.js'
+
+import { BaseParser } from './parser.js'
 
 /**
  * JavaScript/TypeScript parser for GLAD
  */
-export class JsTsParser {
-  /**
-   * @param {GLADContext} context - The shared context
-   */
-  constructor (context) {
-    this.context = context
-  }
-
+export class JsTsParser extends BaseParser {
   /**
    * Generate dependencies from JavaScript/TypeScript source files
    */
@@ -116,72 +109,11 @@ export class JsTsParser {
   }
 
   /**
-   * Normalize the starting path from options.input
-   * @returns {string}
-   */
-  normalizeStartingPath () {
-    let startingPath = this.context.options.input
-    if (startingPath && startingPath !== '.') {
-      startingPath = startingPath.replace(/\/$/, '') // Remove trailing slash if present
-    } else {
-      startingPath = './'
-    }
-    return startingPath.endsWith('/') ? startingPath : startingPath + '/'
-  }
-
-  /**
-   * Create the "Folder to Folder" to "File Container" mapping
-   * @param {string} pathToFile
-   */
-  processFileAsContainer (pathToFile) {
-    const listOfContainer = pathToFile.split('/').filter(token => !stringIsEmpty(token))
-    this.context.graph.rootNode.createContainerMappingBasedOnArray(listOfContainer, pathToFile)
-  }
-
-  /**
    * Common file processing logic for JS/TS files
    * @param {Array} files - Array of TSMorph.SourceFile objects
    * @param {(file: object) => string} getPathFunc - Function to get path from file object
    */
   processFiles (files, getPathFunc) {
-    if (files.length === 0) {
-      return
-    }
-
-    if (!this.context.options.silent) {
-      console.info(`Analyzing ${files.length} files ...`)
-    }
-
-    if (this.context.options.listFiles) {
-      files.forEach((file) => {
-        console.log('\t' + getPathFunc(file))
-      })
-    }
-
-    // Files to Container model instances
-    files.forEach(file => {
-      this.getTargetFiles(file)
-    })
-
-    // Create containers (folder) tree from each file paths
-    this.context.allFiles.forEach(pathToFile => {
-      this.processFileAsContainer(pathToFile)
-    })
-
-    // Create edges between files
-    this.context.allFileImports.forEach(edge => {
-      this.processFilesAsEdges(edge.source, edge.target)
-    })
-
-    // Merge matching .ts & .js
-    this.mergeMatchingTSAndJSFiles()
-  }
-
-  /**
-   * @param {string} pathToFile
-   * @param {string} importFilePath
-   */
-  processFilesAsEdges (pathToFile, importFilePath) {
-    this.context.graph.upsertFileLinkByText(pathToFile, importFilePath)
+    this.processFilesCommon(files, getPathFunc, (file) => this.getTargetFiles(file), true)
   }
 }
