+++ lib/parsers/parserSwift.js
@@ -4,20 +4,14 @@ import path from 'path'
 import Parser from 'tree-sitter'
 import Swift from 'tree-sitter-swift'
 
-import { GLADContext } from '../gladContext.js'
-import { stringIsEmpty } from '../util/strings.js'
+import { matchesExcludePattern } from '../util/strings.js'
+
+import { BaseParser } from './parser.js'
 
 /**
  * Swift parser for GLAD
  */
-export class SwiftParser {
-  /**
-   * @param {GLADContext} context - The shared context
-   */
-  constructor (context) {
-    this.context = context
-  }
-
+export class SwiftParser extends BaseParser {
   /**
    * Helper to detect swift project
    * @param {string} input
@@ -251,53 +245,11 @@ export class SwiftParser {
     // Filter out excluded files
     const filteredFiles = swiftFiles.filter(file => {
       const relativePath = path.relative(startingPath, file)
-      return !this.matchesExcludePattern(relativePath, this.context.options.exclude || [])
+      return !matchesExcludePattern(relativePath, this.context.options.exclude || [])
     })
 
     // Process the Swift files
     this.processSwiftFiles(filteredFiles)
-  }
-
-  /**
-   * Check if a file path matches any exclude pattern
-   * @param {string} filePath - Relative file path to check
-   * @param {string|Array<string>} excludePatterns - Glob pattern(s) to match against
-   * @returns {boolean} True if the file should be excluded
-   */
-  matchesExcludePattern (filePath, excludePatterns) {
-    if (!Array.isArray(excludePatterns)) {
-      excludePatterns = excludePatterns ? [excludePatterns] : []
-    }
-
-    return excludePatterns.some(pattern => {
-      // Simple glob matching implementation
-      // First escape special regex characters including *
-      let regexPattern = pattern.replace(/[.+^${}()|[\]\\*?/]/g, '\\$&')
-
-      // Then convert escaped glob patterns to regex
-      regexPattern = regexPattern
-        .replace(/\\\*\\\*\\\//g, '(.*/)?')  // **/ matches optional path (escaped **/)
-        .replace(/\\\*\\\*/g, '.*')         // ** matches any path (escaped **)
-        .replace(/\\\*/g, '.*')             // * matches any characters including / (escaped *)
-        .replace(/\\\?/g, '[^/]')           // ? matches any single character except / (escaped ?)
-
-      const regex = new RegExp('^' + regexPattern + '$')
-      return regex.test(filePath)
-    })
-  }
-
-  /**
-   * Normalize the starting path from options.input
-   * @returns {string}
-   */
-  normalizeStartingPath () {
-    let startingPath = this.context.options.input
-    if (startingPath && startingPath !== '.') {
-      startingPath = startingPath.replace(/\/$/, '') // Remove trailing slash if present
-    } else {
-      startingPath = './'
-    }
-    return startingPath.endsWith('/') ? startingPath : startingPath + '/'
   }
 
   /**
@@ -312,41 +264,10 @@ export class SwiftParser {
   }
 
   /**
-   * Create the "Folder to Folder" to "File Container" mapping
-   * @param {string} pathToFile
-   */
-  processFileAsContainer (pathToFile) {
-    const listOfContainer = pathToFile.split('/').filter(token => !stringIsEmpty(token))
-    this.context.graph.rootNode.createContainerMappingBasedOnArray(listOfContainer, pathToFile)
-  }
-
-  /**
-   * @param {string} pathToFile
-   * @param {string} importFilePath
-   */
-  processFilesAsEdges (pathToFile, importFilePath) {
-    this.context.graph.upsertFileLinkByText(pathToFile, importFilePath)
-  }
-
-  /**
    * Process Swift files for dependency analysis
    * @param {Array<string>} files - Array of Swift file paths
    */
   processSwiftFiles (files) {
-    if (files.length === 0) {
-      return
-    }
-
-    if (!this.context.options.silent) {
-      console.info(`Analyzing ${files.length} files ...`)
-    }
-
-    if (this.context.options.listFiles) {
-      files.forEach((file) => {
-        console.log('\t' + file)
-      })
-    }
-
     // Map: SymbolName -> FilePath
     const definitions = new Map()
 
@@ -372,10 +293,9 @@ export class SwiftParser {
     })
 
     // Pass 2: Find type usages using AST parsing
-    files.forEach(file => {
+    this.processFilesCommon(files, (file) => file, (file) => {
       try {
         this.context.allFiles.add(file)
-        this.processFileAsContainer(file)
 
         const content = fs.readFileSync(file, 'utf8')
         const tree = this.parseSwiftFile(content)
@@ -399,11 +319,6 @@ export class SwiftParser {
           console.warn(`Warning: Failed to analyze Swift file ${file}: ${error.message}`)
         }
       }
-    })
-
-    // Create edges between files
-    this.context.allFileImports.forEach(edge => {
-      this.processFilesAsEdges(edge.source, edge.target)
     })
   }
