+++ lib/util/strings.js
@@ -26,6 +26,34 @@ export function ensureNotEndingWith (text, textToRemove) {
 }
 
 /**
+ * Check if a file path matches any exclude pattern
+ * @param {string} filePath - Relative file path to check
+ * @param {string|Array<string>} excludePatterns - Glob pattern(s) to match against
+ * @returns {boolean} True if the file should be excluded
+ */
+export function matchesExcludePattern (filePath, excludePatterns) {
+  if (!Array.isArray(excludePatterns)) {
+    excludePatterns = excludePatterns ? [excludePatterns] : []
+  }
+
+  return excludePatterns.some(pattern => {
+    // Simple glob matching implementation
+    // First escape special regex characters including *
+    let regexPattern = pattern.replace(/[.+^${}()|[\]\\*?/]/g, '\\$&')
+
+    // Then convert escaped glob patterns to regex
+    regexPattern = regexPattern
+      .replace(/\\\*\\\*\\\//g, '(.*/)?')  // **/ matches optional path (escaped **/)
+      .replace(/\\\*\\\*/g, '.*')         // ** matches any path (escaped **)
+      .replace(/\\\*/g, '.*')             // * matches any characters including / (escaped *)
+      .replace(/\\\?/g, '[^/]')           // ? matches any single character except / (escaped ?)
+
+    const regex = new RegExp('^' + regexPattern + '$')
+    return regex.test(filePath)
+  })
+}
+
+/**
  * @param {string} text
  * @param {number} quantity
  * @returns {string}
