+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@amazon/glad",
-  "version": "1.3.5",
+  "version": "1.3.6",
   "description": "Generate a Layer Diagram view of your source layers",
   "main": "cli/index.js",
   "author": "jp@vteam.com",
